fix day mappings

+++ db/migrate/20120907144558_create_running_templates.rb
@@ -2,8 +2,8 @@ class CreateRunningTemplates < ActiveRecord::Migration
   def change
     create_table :running_templates do |t|
     	t.string	:name
-    	t.decimal	:distance
-    	t.decimal	:distance_per_week
+    	t.integer	:distance
+    	t.integer	:distance_per_week
     	t.integer   :distance_type
     	t.integer	:level
     	t.integer	:number_of_weeks
