+++ db/migrate/20120907150435_create_running_template_workouts.rb
@@ -2,7 +2,7 @@ class CreateRunningTemplateWorkouts < ActiveRecord::Migration
   def change
     create_table :running_template_workouts do |t|
       t.integer :running_template_week_id
-      t.decimal :total_distance
+      t.integer :total_distance
       t.integer :distance_type
       t.text    :description
       t.timestamps
