+++ db/migrate/20120907151648_create_running_template_workout_legs.rb
@@ -2,7 +2,7 @@ class CreateRunningTemplateWorkoutLegs < ActiveRecord::Migration
   def change
     create_table :running_template_workout_legs do |t|
       t.integer :running_template_workout_id
-      t.decimal  :distance
+      t.integer  :distance
       t.integer :distance_type
       t.integer :pace_type
       t.integer :repetitions
